requirements.txt
einops==0.4.0
GitPython==3.1.26
h5py==3.6.0
matplotlib==3.5.1
numpy==1.22.1
pandas==1.1.4
prettytable==3.0.0
scikit_learn==1.0.2
scipy==1.7.3
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
lightgbm==3.3.5
salesforce-merlion==1.1.1
rrcf==0.4.4
openpyxl==3.0.10
